+++ android-apk/setup_bitrise_token.sh
@@ -1,4 +1,0 @@
-#!/bin/bash
-read -p "Paste Bitrise Personal Access Token: " TOKEN
-echo "$TOKEN" > .bitrise_token
-echo "Token saved"