+++ metro.config.extra.js
@@ -1,6 +1,0 @@
-module.exports = {
-  watchFolders: [],
-  resolver: {
-    blacklistRE: /node_modules\/.*\/(tests|__tests__|examples|docs)\/.*/
-  }
-};